requirements.txt
astral
python-kasa